Initial patch for the metric selection bug
I was trying to use aardvark today and was unable to draw any graphs. I think that the options are being reset either erroneously or too early. This very quick fix allows graphs to be drawn. I don't know whether edge-case functionality is affected by this change. I have done this purely online and so haven't ran any of the tests (although if this causes the tests to fail but an inability to draw graphs does not then I would argue that new tests are needed).